fix(pricing): restore Phase 3 fractional 0.25 floor + 6-decimal precision (#538)
PR4 of Fix the Shit arc.

The legacy Math.max(1, Math.round(...)) clamp in mapModelRow forced
integer 1 and overrode the canonical 0.25 product floor. The integer
DB column (20260514191000) made fractional persistence impossible
anyway. iOS still parsed minimum_charge_credits as Int and rounded
budget math with Int(ceil(...)), quietly truncating every fractional
credit the backend emitted.

Changes:
- supabase/migrations/20260912100000_support_fractional_minimum_charge_credits.sql:
  ALTER minimum_charge_credits TYPE NUMERIC(18, 6), default 0.25,
  backfill Phase 3 lineage (gpt-4o-mini/4o/4.1 family, gpt-5/5.6,
  embeddings) to 0.25, defensive clamp [0, 1000].
- supabase/functions/generate-story/_generation_models.ts:
  - Export mapModelRow for direct testability.
  - Add round6 helper (Math.round(x * 1e6) / 1e6) so mapModelRow and
    computeActualChargeCredits share 6-decimal precision.
  - Replace the legacy integer clamp with
    round6(toNumber(row.minimum_charge_credits, DEFAULT_PRICING.minimumChargeCredits)).
- supabase/functions/generate-story/pricing_test.ts:
  three new test cases guard against regression of the fractional clamp
  (DB row 0.25 preserved, DB null falls back to 0.25, fractional DB
  value round6 to 6 decimals).
- CathedralOSApp/Services/GenerationModelService.swift,
  CoherenceCheckService.swift, GenerationRequestDTO.swift:
  minimumChargeCredits: Int/Int? -> Double/Double? so JSON decoding
  no longer fails on the new NUMERIC payload.
- CathedralOSApp/Features/Projects/ProjectDetailView.swift:
  max(model.minimumChargeCredits, Int(ceil(raw))) -> Double math;
  budget label strips trailing zeros for clean display.

Validation:
- deno check on touched TS files passes.
- deno test -A supabase/functions/generate-story/pricing_test.ts:
  31 passed | 0 failed (3 new fractional cases).
- git diff --check clean.
- iOS Build + Supabase Deploy required (migration + Deno + iOS sites).
- WSL2 has no xcodebuild; CI iOS Build is the build gate.

Refs: docs/pr-plans/2026-09-12-fix-the-shit-pr4-legacy-minimum-and-fractional.md

+++ CathedralOSApp/Services/GenerationRequestDTO.swift
@@ -326,5 +326,5 @@ struct GenerationCostEstimate: Codable {
     /// `true` if the user has enough credits to proceed with generation.
     let allowed: Bool
     /// Minimum charge enforced by the selected model regardless of actual token usage.
-    let minimumChargeCredits: Int
+    let minimumChargeCredits: Double
 }
